% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reform.R
\name{as_sim_age}
\alias{as_sim_age}
\alias{as_sim_age_sex}
\alias{as_sim_all}
\alias{as_sim_sex}
\title{Generating simulation-friendly demographic data}
\usage{
as_sim_age(dat, year0, year1, agp, agl, sex = c("T", "F", "M"),
  ageing_to_dead = FALSE, k = 1, bind = FALSE, multi_task = FALSE)

as_sim_age_sex(dat, year0, year1, agp, agl, sex_ratio = 107, ...)

as_sim_all(dat, year0, year1, sex = c("T", "F", "M"), k = 1)

as_sim_sex(dat, year0, year1, sex_ratio = 107)
}
\arguments{
\item{dat}{}

\item{year0}{}

\item{year1}{}

\item{agp}{}

\item{agl}{}

\item{sex}{"F" female; "M" male; "T" total}

\item{ageing_to_dead}{TRUE if the ageing in the final age group will bring death; otherwise no ageing for them.}

\item{k}{}

\item{bind}{bind the result table (data.table) or not (list)}

\item{multi_task}{cores for parallel computation, FALSE for single tread.}

\item{sex_ratio}{sex ratio at birth, 107 for the global average}
}
\description{
Generating simulation-friendly demographic data
}
\examples{
demo <- fetch_demography()
agp = c(rep(1:16, each=5), rep(16, 21))
agl = c(paste(seq(0, 70, 5), seq(4, 74, 5), sep = "-"), "75+")
res = as_sim_age(dat, 2000, 2005, agp, agl, sex = "T", k = 1, bind = T)
head(res)

res <- as_sim_sex(demo, 2000, 2005)
head(res)

}
