% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{fetch_demography}
\alias{fetch_demography}
\alias{print.pop_raw}
\title{Fetch data from a github repo}
\usage{
fetch_demography(
  folder = "ByCountry/United Kingdom",
  repo = "TimeWz667/pop4modelling",
  branch = "master"
)

\method{print}{pop_raw}(obj)
}
\arguments{
\item{folder}{folder of data within a repo}

\item{repo}{name of the repo}

\item{branch}{selected branch}
}
\value{
a 'pop_raw' object
}
\description{
Fetch data from a github repo
}
\examples{
demo <- fetch_demography(folder = "ByCountry/United Kingdom")
print(demo)

}
