% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterise.R
\name{as_pars_all}
\alias{as_pars_all}
\alias{as_pars_sex}
\title{Title}
\usage{
as_pars_all(sim, pkg = c("odin", "deSolve"))

as_pars_sex(sim, pkg = c("odin", "deSolve"))
}
\arguments{
\item{sim}{}

\item{pkg}{}
}
\value{

}
\description{
Title
}
